chore: remove consignFromCollectorProfileMyCollectionRoutes add routes to consignRoutes
